requirement.txt
pywasm3==0.5.0